Base project setup in ReactJS
