+++ web/src/components/app/App.css
@@ -1,0 +1,4 @@
+.App {
+  width: 100%;
+  height: 100%;
+}
