+++ web/src/components/app/App.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import './App.css';
+import Header from './header/Header';
+
+function App() {
+  return (
+    <div className="App">
+      <Header>
+
+      </Header>
+    </div>
+  );
+}
+
+export default App;
