+++ web/src/components/app/header/Header.css
@@ -1,0 +1,18 @@
+.home-banner {
+  background-image: url('../../../assets/images/banner.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
+  text-align: center;
+  height: 45%;
+  padding: 0 !important;
+}
+
+.filter {
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.3);
+}
+
+.welcome-title {
+  font-family: 'Source Sans Pro', 'Helvetica Neue', sans-serif;
+}
